removed mmtk tests and fixed test in test_configurator, removed tutu projector test which doesn't exist anymore
